Add send_deltas param to shape definition
Phoenix.LiveView Stream integration requires the full row to perform an
update, not just the changed values. This adds a parameter to the shape
definition that turns off the column diffing for updates. Since we were
just dropping unchanged values this is a minor change to the wiring.
Most of the work here is adding the configuration.